Refactored dust transition particle tests from 1.16.5 particle tests
